Secured game object
So I made the variable game local in both game and client. Here's how it
works:

There is a global variable called gameStarter with the function "start".

When the client starts, it calls this function while making its game
object equal to the return value of the start function.

The start function returns a pointer to its local game object and then
sets the "start" function of the gameStarter to null.

Now, there is a connection between client and game and it isn't global.