Server Connection
Altered the code in the game and server so that the server could run the
game as well. When the global variable is present, the game exports the
startGame object to the server.

I wanted to set the object as a function but because the object being
returned to the server isn't technically the actual startGame object
setting "startGame" to undefined didn't work on the server. I also tried
"this = undefined" but that didn't work either. What ended up working
was having a function within the startGame object called "start" and
setting that to undefiend in the start function by saying "this.start =
undefined".