Color Performance
So, one of the issues I noticed with the coloring of the images, was
that the coloring was happening every frame. This means that the image
was being drawn, then the image was being manipulated pixel by pixel to
color it, every frame. This caused a performance drop. To fix this, I
made the coloring only happen on the first frame and any time the color
changed. This way, performance is much better. One of the things I
realized while doing this, is that if you try to do this the first
frame, it won't work because the image hasn't loaded yet. So, I also had
to add a property called ready to each object that's drawable so that
when the image is finished loading it will be drawn. Also, on the first
frame, the image has to be drawn at (0,0) because if the image is off
frame, the client doesn't know to redraw those pixels and therfore when
they come into the frame, they are black. The other choice was to have a
trigger that causes the recoloring so that the image comes into view. I
haven't tested this with images that have more transparent pixels and
juging by the transparent pixels that show up in the player, I have a
feeling they won't be transparent. We'll see.