Color Transparency
So, turns out transparency didn't work as intended but I found a
solution. Now, there are two seperate textures for each drawable, an
original and a draw texture. When applying a color, a new canvas is made
where the image is drawn, the color is applied and then the result is
converted back into an image. Suprisingly, the performance isn't that
effected by it. In this example, the shot color is directly tied to the
player color so that when the player color changes, the shot color
matches (For example). The red value of the player increase until 255
and then is reset at 0.